requirements.txt
openai>=0.28
streamlit>=1.0.0
Pillow>=9.0.0
python-dotenv>=0.19.0 